testing filter

+++ hub.py
@@ -26,6 +26,48 @@ cache.setblocking(0)
 cache.bind(CACHEPATH)
 CACHEFD=cache.fileno()
 
+sys.dont_write_bytecode=True
+import config, udpmsg4
+config.mtime=os.path.getmtime('config.py')
+
+def filter(kvps):
+    if kvps==0:
+        return 0
+    if config.accept!={}:
+        for key in config.accept.keys():
+            if config.accept[key]==None:
+                if not key in kvps.keys():
+                    for this_key in kvps.keys():
+                        if not this_key in config.accept:
+                            os.write(2,'hub.py: '+CLIENT+' KEY ['+this_key+'] NOT IN ACCEPT\n')
+                            return 0
+            if config.accept[key]!=None:
+                if key in kvps:
+                    if config.accept[key]!=kvps[key]:
+                        os.write(2,'hub.py: '+CLIENT+' KEY/VALUE ['+key+'='+config.accept[key]+'] != ['+key+'='+kvps[key]+']\n')
+                        return 0
+        if None in config.accept.keys():
+            if not config.accept[None] in kvps.values():
+                os.write(2,'hub.py: '+CLIENT+' VALUE ['+config.accept[None]+'] NOT IN PACKET\n')
+                return 0
+
+    if config.reject!={}:
+        for key in config.reject.keys():
+            if config.reject[key]==None:
+                if key in kvps.keys():
+                    os.write(2,'hub.py: '+CLIENT+' KEY ['+key+'] IN REJECT\n')
+                    return 0
+            if config.reject[key]!=None:
+                if key in kvps.keys():
+                    if config.reject[key]==kvps[key]:
+                        os.write(2,'hub.py: '+CLIENT+' KEY/VALUE ['+key+'='+config.reject[key]+'] == ['+key+'='+kvps[key]+']\n')
+                        return 0
+        if None in config.reject.keys():
+            if config.reject[None] in kvps.values():
+                os.write(2,'hub.py: '+CLIENT+' VALUE ['+config.reject[None]+'] IN PACKET\n')
+                return 0
+    return kvps
+
 LOOP_TIME=0
 CLIENT_QUEUE=[]
 SOCKET_QUEUE={}
@@ -33,6 +75,10 @@ SENDSOCKETS={}
 RECVSOCKETS={}
 REMOTE_QUEUE='[]'
 while 1:
+    if config.mtime!=os.path.getmtime('config.py'):
+        config.mtime=os.path.getmtime('config.py')
+        reload(config)
+
     TIME=time()
     TIMEOUT=1.0/(1+LOOP_TIME)
     REMOTE_SOCKS=[]
@@ -56,7 +102,7 @@ while 1:
         if ex.errno == 24:
             os.write(2,'hub.py: '+CLIENT+' failed to read pids\n') 
     MAX_QUEUE=128*(len(REMOTE_SOCKS)+1)
-    os.write(2,'hub.py: '+CLIENT+' MAX_QUEUE=['+str(MAX_QUEUE)+'] TIMEOUT=['+str(TIMEOUT)+'] REMOTE_QUEUE='+REMOTE_QUEUE+'\n')
+    os.write(2,'hub.py: '+CLIENT+' MAX_QUEUE=['+str(MAX_QUEUE)+'] TIMEOUT=['+str(TIMEOUT)+'] CLIENT_QUEUE=['+str(len(CLIENT_QUEUE))+'] REMOTE_QUEUE='+REMOTE_QUEUE+'\n')
 
     if len(POLL([0],[],[],TIMEOUT)[0])!=0:
         try:
@@ -114,26 +160,29 @@ while 1:
                     except:
                         pass
             if cache_query=='\x00':
-                os.write(2,'hub.py: '+CLIENT+' successful read from client\n')
-                for REMOTE in REMOTE_SOCKS:
-                    if not REMOTE in SOCKET_QUEUE:
-                        try:
-                            SENDSOCKETS[REMOTE]=socket.socket(socket.AF_UNIX,socket.SOCK_DGRAM)
-                            SENDSOCKETS[REMOTE].setblocking(0)
+                if filter(udpmsg4.unframe(packet))!=0:
+                    os.write(2,'hub.py: '+CLIENT+' successful read from client\n')
+                    for REMOTE in REMOTE_SOCKS:
+                        if not REMOTE in SOCKET_QUEUE:
                             try:
-                                os.remove(HUBDIR+'/send/'+PID+'TO'+REMOTE)
-                            except:
-                                pass
-                            SENDSOCKETS[REMOTE].bind((HUBDIR+'/send/'+PID+'TO'+REMOTE))
-                            SOCKET_QUEUE[REMOTE]=collections.deque([],MAX_QUEUE)
-                        except socket.error, ex:
-                            if ex.errno == 24:
-                                os.write(2,'hub.py: '+CLIENT+' failed to create sendsocket for '+REMOTE+'\n')
-                    if REMOTE in SOCKET_QUEUE:
-                        SOCKET_QUEUE[REMOTE]+=[packet]
+                                SENDSOCKETS[REMOTE]=socket.socket(socket.AF_UNIX,socket.SOCK_DGRAM)
+                                SENDSOCKETS[REMOTE].setblocking(0)
+                                try:
+                                    os.remove(HUBDIR+'/send/'+PID+'TO'+REMOTE)
+                                except:
+                                    pass
+                                SENDSOCKETS[REMOTE].bind((HUBDIR+'/send/'+PID+'TO'+REMOTE))
+                                SOCKET_QUEUE[REMOTE]=collections.deque([],MAX_QUEUE)
+                            except socket.error, ex:
+                                if ex.errno == 24:
+                                    os.write(2,'hub.py: '+CLIENT+' failed to create sendsocket for '+REMOTE+'\n')
+                        if REMOTE in SOCKET_QUEUE:
+                            SOCKET_QUEUE[REMOTE]+=[packet]
 
     for SOCKET in RECVSOCKETS:
         while POLL([RECVSOCKETS[SOCKET].fileno()],[],[],0)[0]!=[]:
+            if len(CLIENT_QUEUE)<=MAX_QUEUE:
+                CLIENT_QUEUE=collections.deque(CLIENT_QUEUE,MAX_QUEUE)
             try:
                 packet_length=0
                 packet,REMOTE=RECVSOCKETS[SOCKET].recvfrom(65536)
@@ -144,31 +193,24 @@ while 1:
                 os.write(2,'hub.py: '+CLIENT+' connection to '+REMOTE[len(HUBDIR)+6::].split('TO')[0]+' died\n')
             elif packet_length!=len(packet[2::]):
                 os.write(2,'hub.py: '+CLIENT+' rejected protocol error from '+REMOTE[len(HUBDIR)+6::].split('TO')[0]+'\n')
-            else:
+            if filter(udpmsg4.unframe(packet))!=0:
                 CLIENT_QUEUE+=[packet]
-                os.write(2,'hub.py: '+CLIENT+' successful read from '+REMOTE[len(HUBDIR)+6::].split('TO')[0]+'\n')
-                if len(CLIENT_QUEUE)<=MAX_QUEUE:
-                    CLIENT_QUEUE=collections.deque(CLIENT_QUEUE,MAX_QUEUE)
-
-    while len(CLIENT_QUEUE)!=0:
-        write_length=0
-        packet=CLIENT_QUEUE[0]
-        packet_length=len(packet)
-        while packet_length!=write_length:
-            try:
-                write_length=os.write(1,packet[write_length::])
-            except:
-                break
-        if packet_length==write_length:
-            CLIENT_QUEUE.popleft()
-            os.write(2,'hub.py: '+CLIENT+' successful write to client\n')
-        elif write_length>0:
-            CLIENT_QUEUE[0]=packet[write_length::]
-            os.write(2,'hub.py: '+CLIENT+' could not write complete packet to client\n')
-            break
-        elif write_length==0:
-            os.write(2,'hub.py: '+CLIENT+' failed to write to client\n')
-            break
+                write_length=0
+                packet=CLIENT_QUEUE[0]
+                packet_length=len(packet)
+                while packet_length!=write_length:
+                    try:
+                        write_length=os.write(1,packet[write_length::])
+                    except:
+                        break
+                if packet_length==write_length:
+                    CLIENT_QUEUE.popleft()
+                    os.write(2,'hub.py: '+CLIENT+' successful write to client\n')
+                elif write_length>0:
+                    CLIENT_QUEUE[0]=[packet[write_length::]]
+                    os.write(2,'hub.py: '+CLIENT+' could not write complete packet to client\n')
+                elif write_length==0:
+                    os.write(2,'hub.py: '+CLIENT+' failed to write to client\n')
 
     if SOCKET_QUEUE!={}:
         DEQUE=[]
@@ -213,4 +255,5 @@ while 1:
                 pass
     else:
         REMOTE_QUEUE='[]'
+
     LOOP_TIME=time()-TIME
